Add PropType definitions invoke close callback when button is clicked break ProgressBar into its own file
